feat(itemeditor): 完善组件包装器并新增 UI 组件
- 新增 Equippable、Tool Data Component 包装器
- 新增 HolderSetBlockEditorDialog、HolderSetEntityTypeEditorDialog
- 重构 DamageResistant、AttributeModifiers、Repairable 包装器
- 新增 ScrollState 工具及英文翻译
- 完善大量 Data Component 中文翻译条目
- RichTextEditor 支持主题色预览
- 修复 ClientMisc.asTranslateKey 后缀拼接
- 重命名 vanillaTooltip.kt 为首字母大写

+++ common/src/main/kotlin/moe/forpleuvoir/hiirosakura/ui/widget/radialmenu/RadialMenuGeometry.kt
@@ -138,7 +138,7 @@ fun getSectorPath(
 
     // 半径 r 处的单侧间隙角（度），使间隙线性宽度恒为 gap
     fun gapHalfAngle(radius: Float): Float =
-        if (gap > 0f && radius > 0f) Math.toDegrees((gap / (2.0 * radius)).toDouble()).toFloat() else 0f
+        if (gap > 0f && radius > 0f) Math.toDegrees((gap / (2.0 * radius))).toFloat() else 0f
 
     // 边曲线上的点（相对菜单中心）：startEdge 为起始边，否则为结束边
     fun edgePoint(startEdge: Boolean, radius: Float): Offset {
